+++ src/herbs2gql.js
@@ -5,4 +5,5 @@ module.exports = {
     usecase2query: require('./usecase2query'),
     usecase2subscription: require('./usecase2subscription'),
     defaultResolver: require('./defaultResolver'),
+    args2request : require('./args2request'),
 }
